refactor(instrumentations): create constants file
