+++ packages/web-sdk/src/instrumentations/instrumentationConstants.ts
@@ -1,0 +1,1 @@
+export const NAVIGATION_ID_STORAGE_KEY = 'com.grafana.faro.lastNavigationId';
