+++ packages/web-sdk/src/instrumentations/performance/index.ts
@@ -1,2 +1,1 @@
 export { PerformanceInstrumentation } from './instrumentation';
-export { NAVIGATION_ID_STORAGE_KEY } from './performanceConstants';
